+++ lib/src/async/stream_ext.dart
@@ -1,12 +1,19 @@
 import 'dart:async';
 
+import 'package:an_lifecycle_cancellable/an_lifecycle_cancellable.dart';
+import 'package:cancellable/cancellable.dart';
+
 extension StreamDoneTimeoutExt<T> on Stream<T> {
   /// 设定stream 的 done 超时
+  /// * [duration] 超时时间
+  /// * [onTimeout] 超时时返回的值
+  /// * [cancelOnError] 发生错误时 取消
   Stream<T> timeoutDone(Duration duration,
       {void Function(StreamSink<T> sink)? onTimeout,
       bool? cancelOnError = true}) {
-    StreamController<T> controller =
-        isBroadcast ? StreamController.broadcast() : StreamController();
+    StreamController<T> controller = isBroadcast
+        ? StreamController.broadcast(sync: true)
+        : StreamController(sync: true);
     Timer? time;
 
     void onDone() {
@@ -26,7 +33,9 @@ extension StreamDoneTimeoutExt<T> on Stream<T> {
       time = Timer(duration, () {
         if (!controller.isClosed) {
           onTimeout?.call(controller.sink);
-          controller.close();
+          if (!controller.isClosed) {
+            controller.close();
+          }
         }
       });
       controller.onCancel = () {
@@ -41,12 +50,18 @@ extension StreamDoneTimeoutExt<T> on Stream<T> {
 
 extension StreamToolsExt<T> on Stream<T> {
   /// 每次触发数据时
+  /// * [onData] 回调
   Stream<T> onData(void Function(T event) onData) => map((event) {
         onData(event);
         return event;
       });
 
-  /// 重复上一个
+  /// 重复上一个，对源stream会一直保持订阅，永远不会取消，建议前向使用[bindCancellable]或者[bindLifecycle]，来自动解除订阅
+  /// * [repeatTimeout] 重复上一个的超时时间，如果[onTimeout]和[onRepeatTimeout]都不设置时，表示超时后清除重复数据
+  /// * [onTimeout] 超时时返回的值,仅支持非[null]的值
+  /// * [onRepeatTimeout] 超时时返回的值,允许[null]值
+  /// * [repeatError] 是否重复错误
+  /// * [broadcast] 是否广播
   Stream<T> repeatLatest(
       {Duration? repeatTimeout,
       T? onTimeout,
@@ -54,7 +69,7 @@ extension StreamToolsExt<T> on Stream<T> {
       bool repeatError = false,
       bool? broadcast}) {
     var done = false;
-    T? latest;
+    _RepeatEntry<T>? latest;
     Object? cacheError;
     StackTrace? cacheStackTrace;
 
@@ -72,21 +87,30 @@ extension StreamToolsExt<T> on Stream<T> {
 
     void Function(T value) setLatest = (value) {
       cleanCache();
-      latest = value;
+      latest = _RepeatEntry(value);
     };
 
     Timer? timer;
     if (repeatTimeout != null && repeatTimeout > Duration.zero) {
-      void Function() timeoutCallBack = () => latest = onTimeout;
+      void Function() timeoutCallBack = cleanCache;
+      if (onTimeout != null) {
+        timeoutCallBack = () {
+          cleanCache();
+          latest = _RepeatEntry(onTimeout);
+        };
+      }
 
       if (onRepeatTimeout != null) {
-        timeoutCallBack = () => latest = onRepeatTimeout();
+        timeoutCallBack = () {
+          cleanCache();
+          latest = _RepeatEntry(onRepeatTimeout());
+        };
       }
 
       setLatest = (value) {
         timer?.cancel();
         cleanCache();
-        latest = value;
+        latest = _RepeatEntry(value);
         if (!done) {
           timer = Timer(repeatTimeout, timeoutCallBack);
         }
@@ -101,7 +125,7 @@ extension StreamToolsExt<T> on Stream<T> {
       var latestValue = latest;
       if (latestValue != null) {
         if (!controller.isClosed) {
-          controller.add(latestValue);
+          controller.add(latestValue.value);
         }
       } else if (cacheError != null) {
         if (!controller.isClosed) {
@@ -150,4 +174,10 @@ extension StreamToolsExt<T> on Stream<T> {
       controller.onCancel = () => currentListeners.remove(controller);
     }, isBroadcast: isBroadcast_);
   }
+}
+
+class _RepeatEntry<T> {
+  final T value;
+
+  _RepeatEntry(this.value);
 }